requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via django
cachetools==5.3.3
    # via google-auth
certifi==2024.2.2
    # via requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
cryptography==42.0.7
    # via pyopenssl
django==5.0.6
    # via
    #   -r requirements.in
    #   django-model-utils
    #   djangorestframework
django-model-utils==4.5.1
    # via -r requirements.in
djangorestframework==3.15.1
    # via -r requirements.in
google-api-core==2.19.0
    # via google-api-python-client
google-api-python-client==2.131.0
    # via pydrive2
google-auth==2.29.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
google-auth-httplib2==0.2.0
    # via google-api-python-client
googleapis-common-protos==1.63.0
    # via google-api-core
gunicorn==22.0.0
    # via -r requirements.in
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
    #   oauth2client
idna==3.7
    # via requests
oauth2client==4.1.3
    # via pydrive2
packaging==24.0
    # via gunicorn
proto-plus==1.23.0
    # via google-api-core
protobuf==4.25.3
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.6.0
    # via
    #   oauth2client
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via
    #   google-auth
    #   oauth2client
pycparser==2.22
    # via cffi
pydrive2==1.19.0
    # via -r requirements.in
pyopenssl==24.1.0
    # via pydrive2
pyparsing==3.1.2
    # via httplib2
pyyaml==6.0.1
    # via pydrive2
requests==2.32.3
    # via google-api-core
rsa==4.9
    # via
    #   google-auth
    #   oauth2client
six==1.16.0
    # via oauth2client
sqlparse==0.5.0
    # via django
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.2.1
    # via requests
whitenoise==6.6.0
    # via -r requirements.in
